requirements.txt
pytest>=7.0.0
appium-python-client>=2.0.0
